change comments js to jsx, split file into own modules, convert some modules to jsx already
