+++ euth/contrib/static/js/app.js
@@ -1,4 +1,4 @@
-var ReactComments = require('../../../comments/static/comments/react_comments.js')
+var ReactComments = require('../../../comments/static/comments/react_comments.jsx')
 var ReactRatings = require('../../../ratings/static/ratings/react_ratings.jsx')
 var ReactLanguageSwitch = require('../../../dashboard/static/language_switch/react_language_switch.jsx')
